snapintime: In order to simplify things a bit, I've decided to change it so that week and daily clearing of snapshots aren't looking forever into the future. Rather, it's going to behave as if the computer was always on and if it isn't, it'll be taken care of the next time around.
EG: daily cleans up every third day's worth of snaps. If you don't turn your computer on for a few days, then there will be some days that don't get cleared of snaps. That's OK. When the weekly cleanup happens, it'll automatically take care of those days. And, since weekly is a rolling weekly, it should have a pretty good chance of catching it. If your computer is off even longer, it'll be caught on the quarterly. And that should be good enough.

unittests:
Working on making sure the edge cases work correctly.